refactor: impl task_form widget and use it in common
